+++ src/ua/my/Main.java
@@ -1,76 +1,30 @@
 package ua.my;
+
+import ua.my.manager.Manager;
+import ua.my.threads.*;
 /**
  * Simple program to demonstrate inter-thread communication
  * (wait(), join() and notify()) in Java
  */
-import java.util.Scanner;
 public class Main
 {
-    public static void main(String[] args) throws InterruptedException {
+    public static void main(String[] args) {
 
-        Manager man = new Manager();
 
-        // Create a thread object that calls man.produce()
-        Thread t1 = new Thread(new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    man.produce();
-                } catch(InterruptedException e) {
-                    e.printStackTrace();
-                }
-            }
-        });
 
-        // Create another thread object that calls
-        // man.consume()
-        Thread t2 = new Thread(new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    man.consume();
-                } catch(InterruptedException e) {
-                    e.printStackTrace();
-                }
-            }
-        });
-        // Starting both threads
-        t1.start();
-        t2.start();
-    }
+        // Create a thread object that calls man.waiter()
+        Thread threadWaiter1 = new Thread(new ThreadWaiter(), "Waiter 1");
 
-    /**
-     * Manager class with produce() and consume() methods.
-     */
-    public static class Manager {
-        // Prints a string and waits for consume()
-        public void produce()throws InterruptedException {
-            // synchronized block ensures only one thread running at a time.
-            synchronized(this) {
-                System.out.println("producer thread running");
-                // releases the lock on shared resource
-                wait();
-                // and waits till some other method invokes notify().
-                System.out.println("Resumed");
-            }
-        }
-        // Sleeps for some time and waits for a key press. After key
-        // is pressed, it notifies produce().
-        public void consume()throws InterruptedException {
-            // this makes the produce thread to run first.
-            Thread.sleep(1000);
-            Scanner s = new Scanner(System.in);
-            // synchronized block ensures only one thread
-            // running at a time.
-            synchronized(this) {
-                System.out.println("Waiting for return key.");
-                s.nextLine();
-                System.out.println("Return key pressed");
-                // notifies the produce thread that it can wake up.
-                notify();
-                // Sleep
-                Thread.sleep(1000);
-            }
-        }
+        // Create another thread object that calls man.waiter() again
+        Thread threadWaiter2 = new Thread(new ThreadWaiter(), "Waiter 2");
+
+        // Create another thread object that calls man.notifier()
+        Thread threadNotifier = new ThreadNotifier();
+        threadNotifier.setName("Notifier");
+
+        // Starting all threads
+        threadWaiter1.start();
+        threadWaiter2.start();
+        threadNotifier.start();
     }
 }
