+++ src/ua/my/threads/ThreadWaiter.java
@@ -1,0 +1,14 @@
+package ua.my.threads;
+
+import ua.my.manager.Manager;
+
+public class ThreadWaiter implements Runnable {
+    public void run() {
+        Manager man = new Manager();
+        try {
+            man.waiter();
+        } catch(InterruptedException e) {
+            e.printStackTrace();
+        }
+    }
+}
